client/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2
django-uniauth==1.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
lxml==4.6.3
mypy-extensions==0.4.3
pathspec==0.8.1
PyJWT==1.7.1
python-cas==1.5.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
